+++ src/Algorithm/Luhn/LuhnValidator.cs
@@ -95,13 +95,10 @@ public static class LuhnValidator
     public static bool IsValidLuhnCheckDigit(this char checkDigit, ReadOnlySpan<char> number)
     {
         checkDigit.ThrowIfNotDigit();
-
         var trimmedNumber = number.ValidateAndTrimNumber();
-        Span<char> luhnNumber = stackalloc char[trimmedNumber.Length + 1];
-        trimmedNumber.CopyTo(luhnNumber[..^1]);
-        luhnNumber[^1] = checkDigit;
-        ReadOnlySpan<char> readOnlyLuhnNumber = luhnNumber;
-        return readOnlyLuhnNumber.DoubleEverySecondDigit(true) == 0;
+        return trimmedNumber.CreateNumberWithCheckDigitAndValidate(
+            checkDigit,
+            n => n.DoubleEverySecondDigit(true) == 0);
     }
 #endif