+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Moved `ToUnsignedIntegerDigit` method from `LuhnDotNetCore` to `CharExtensions` class.
+
+### Fixed
+- Fixed StackOverflowException in validation methods when the input is too long.
+
 ## [2.0.0] - 2025-06-29
 ### Added
 - Added internal `LuhnAlgorithm` class to encapsulate Luhn algorithm related methods.