+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Moved `ToUnsignedIntegerDigit` method from `LuhnDotNetCore` to `CharExtensions` class.
+- Optimized `IsDigits` method in `ReadOnlySpanExtensions` class.
+- Optimized `ComputeNumberWithCheckDigit` method in `ReadOnlySpanExtensions` class.
+- Renamed `CreateNumberWithCheckDigit` method to `CreateNumberWithCheckDigitAndValidate` in `ReadOnlySpanExtensions` class.
+
+### Fixed
+- Fixed StackOverflowException in validation methods when the input is too long.
+- Fixed .NET Standard 2.0 issues with Damm algorithm and Mod11 Ascending Weights algorithm implementations.
+
 ## [2.0.0] - 2025-06-29
 ### Added
 - Added internal `LuhnAlgorithm` class to encapsulate Luhn algorithm related methods.