+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.1] - 2025-06-22
 
 ### Changed
 - Moved `ToUnsignedIntegerDigit` method from `LuhnDotNetCore` to `CharExtensions` class.
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed StackOverflowException in validation methods when the input is too long.
 - Fixed .NET Standard 2.0 issues with Damm algorithm and Mod11 Ascending Weights algorithm implementations.
 
-## [2.0.0] - 2025-06-29
+## [2.0.0] - 2025-06-19
 ### Added
 - Added internal `LuhnAlgorithm` class to encapsulate Luhn algorithm related methods.
 - Added public `LuhnCalculator` class to encapsulate Luhn calculation methods.
@@ -120,6 +120,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added initial version of LuhnDotNet
 
+[2.0.1]: https://github.com/shinji-san/LuhnDotNet/compare/v2.0.0..v2.0.1
 [2.0.0]: https://github.com/shinji-san/LuhnDotNet/compare/v1.3.0..v2.0.0
 [1.3.0]: https://github.com/shinji-san/LuhnDotNet/compare/v1.2.0..v1.3.0
 [1.2.0]: https://github.com/shinji-san/LuhnDotNet/compare/v1.1.0..v1.2.0